feat(test): require every hover_syntax to name only real words (Phase 5.2)
Structural-constraint ledger item 10 (convention -> structure): a second
consistency test requires every `Symbol` a `hover_syntax` example names to
resolve — after alias canonicalization — to a Coreword-registry entry. This
covers operators (`+`), modifiers (`. ,,`), casts (`>CF`), and `@`-module words
(`MATH@SQRT`), all of which canonicalize to registry entries. It catches a doc
example that references a removed or misspelled word and forces every example to
be a concrete runnable snippet rather than a schematic one.

Writing it exposed two more non-runnable examples that used metavariable
placeholders, now made concrete and verified to run:
- COMPARE-WITHIN `a b 64 COMPARE-WITHIN` -> `1/3 1/2 64 COMPARE-WITHIN` (-> -1/1)
- FLOW `xs ~ { ... } MAP` -> `[ 1 2 3 ] ~ { [ 2 ] * } MAP` (-> [ 2 4 6 ])

Both hover_syntax consistency tests (item 9 tokenizes, item 10 resolves) move
into a sibling `builtin_word_details_tests.rs`, co-locating the two checks and
keeping `builtin_word_details.rs` within the §14.1 budget (478 lines). SKILL.md
regenerated for the two example fixes; the ledger records item 10 done, with
full-execution of non-fragment snippets deferred as item 10b.
