+++ docs/provenance/source-root.txt
@@ -1,1 +1,1 @@
-sha256:a5f22c8f97a429c30186c196879712f90787dc48c93f7c1063861124730d33cd
+sha256:7398c4db922f906386f8e1d22dd6628805c77cc29e8257c9e4435a9958a61829
